Compiles but segfaults (surprise!)
