+++ CMakeLists.txt
@@ -12,6 +12,12 @@ find_package(PkgConfig)
 
 find_package(OpenCV REQUIRED)
 
+find_package(Qt5 COMPONENTS Widgets REQUIRED)
+
+set(CMAKE_AUTOMOC ON)
+set(CMAKE_AUTORCC ON)
+set(CMAKE_AUTOUIC ON)
+
 pkg_check_modules(LIBCAMERA REQUIRED IMPORTED_TARGET libcamera)
 message(STATUS "libcamera library found:")
 message(STATUS "    version: ${LIBCAMERA_VERSION}")
@@ -19,7 +25,15 @@ message(STATUS "    libraries: ${LIBCAMERA_LINK_LIBRARIES}")
 message(STATUS "    include path: ${LIBCAMERA_INCLUDE_DIRS}")
 
 include_directories(${CMAKE_SOURCE_DIR} ${LIBCAMERA_INCLUDE_DIRS} ${OPENCV_INCLUDE_DIRS})
+include(GNUInstallDirs)
 
-add_executable(simple-cam simple-cam.cpp)
+add_executable(webcam-viewer
+  window.cpp
+  main.cpp
+  libcam2opencv.cpp
+)
 
-target_link_libraries(simple-cam PkgConfig::LIBCAMERA ${OpenCV_LIBS})
+target_link_libraries(webcam-viewer PkgConfig::LIBCAMERA)
+target_link_libraries(webcam-viewer Qt5::Widgets)
+target_link_libraries(webcam-viewer qwt-qt5)
+target_link_libraries(webcam-viewer ${OpenCV_LIBS})
